Renamed Template to Base classes; moved NormalNoise to definition
