+++ tests/integration_tests/test_multiple_sensors.py
@@ -17,10 +17,10 @@ import unittest
 import numpy as np
 
 # local imports
-from probeye.definition.forward_model import ForwardModelTemplate
+from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor, PositionSensor
 from probeye.definition.inference_problem import InferenceProblem
-from probeye.inference.taralli_.noise_models import NormalNoise
+from probeye.definition.noise_model import NormalNoise
 from probeye.inference.taralli_.solver import run_taralli_solver
 from probeye.inference.taralli_.postprocessing import run_taralli_postprocessing
 
@@ -85,7 +85,7 @@ class TestProblem(unittest.TestCase):
         #                       Define the Forward Model                       #
         # ==================================================================== #
 
-        class LinearModel(ForwardModelTemplate):
+        class LinearModel(ForwardModelBase):
             def __call__(self, inp):
                 t = inp['time']
                 A = inp['A']
