+++ tests/integration_tests/test_two_models.py
@@ -18,10 +18,10 @@ import numpy as np
 import matplotlib.pyplot as plt
 
 # local imports
-from probeye.definition.forward_model import ForwardModelTemplate
+from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
 from probeye.definition.inference_problem import InferenceProblem
-from probeye.inference.taralli_.noise_models import NormalNoise
+from probeye.definition.noise_model import NormalNoise
 from probeye.inference.taralli_.solver import run_taralli_solver
 from probeye.inference.taralli_.postprocessing import run_taralli_postprocessing
 
@@ -80,7 +80,7 @@ class TestProblem(unittest.TestCase):
         #                      Define the Forward Models                       #
         # ==================================================================== #
 
-        class LinearModel(ForwardModelTemplate):
+        class LinearModel(ForwardModelBase):
             def __call__(self, inp):
                 x = inp['x']
                 a = inp['a']
@@ -90,7 +90,7 @@ class TestProblem(unittest.TestCase):
                     response[os.name] = a * x + b
                 return response
 
-        class QuadraticModel(ForwardModelTemplate):
+        class QuadraticModel(ForwardModelBase):
             def __call__(self, inp):
                 x = inp['x']
                 alpha = inp['alpha']
